+++ yast/requests.py
@@ -100,6 +100,9 @@ class Request(Mapping):
 
     @property
     def session(self):
+        assert "session" in self._scope, (
+            "`SessionMiddleware` must be " "installed to access request.session"
+        )
         return self._scope["session"]
 
     def url_for(self, name: str, **path_params: typing.Any) -> URL: